requirements.txt
pandas>=2.0.0
meteostat>=1.6.7
google-cloud-bigquery>=3.12.0